+++ mercury/explainability/__init__.py
@@ -1,14 +1,22 @@
 __version__ = '1.2.3'
 
-from .explainers.counter_fact_basic import CounterFactualExplainerBasic
-from .explainers.shuffle_importance import ShuffleImportanceExplainer
-from .explainers.explainer import MercuryExplainer
-from .explainers.partial_dependence import PartialDependenceExplainer
-from .explanations.anchors import AnchorsWithImportanceExplanation
-from .explanations.counter_factual import CounterfactualWithImportanceExplanation
-
 from .explainers import ALEExplainer
 from .explainers import AnchorsWithImportanceExplainer
 from .explainers import CounterfactualExplainer, CounterfactualProtoExplainer
+
+from .explainers.clustering_tree_explainer import ClusteringTreeExplainer
+from .explainers.counter_fact_basic import CounterFactualExplainerBasic
+from .explainers.explainer import MercuryExplainer
+from .explainers.monotonicity import MonotonicityExplainer
+from .explainers.partial_dependence import PartialDependenceExplainer
+from .explainers.shuffle_importance import ShuffleImportanceExplainer
+
+from .explanations import AnchorsWithImportanceExplanation
+from .explanations import ClusteringTreeExplanation
+from .explanations import CounterfactualBasicExplanation
+from .explanations import CounterfactualWithImportanceExplanation
+from .explanations import MonotonicityExplanation
+from .explanations import PartialDependenceExplanation
+from .explanations import FeatureImportanceExplanation
 
 from .create_tutorials import create_tutorials
